fix(silentpayments): make Commands::ToSilentPayment deterministic
Before this commit the code was assuming the first output of the
transaction was the one to replace. But this could have been the change
output, making the amount used to create the silent payment output
non-deterministic. To fix this and avoid having to check if the output
can be sign with the Internal or External Keychain, a new "amount"
parameter is added to the command.
The assumption is the PSBT received is going to be a 1In-2Out or a
1In-1Out kind of transaction. So the amount parameter is going to be used
to select the output matching that same amount and replace it with
the silent payment output.
There is still a low probability of the change output being replaced
instead of the intended one, in cases where the 1In-2Out tx has two
outputs with the exact same amount.
As the example is using an address of the same wallet, there is
no risk of losing funds, but in the above case the change output would
end up being received to that external address. To avoid this semantic
issue, the address could be generated using Keychain::Internal.