fix(silentpayments): ensure exhaustion of outputs_to_check before updating tweak
Previous code had a bug that caused the miss of some spouts sent to the
wallet. By recomputing the tweak used for scanning each time a new spout
was found, the code was ignoring all the remaining combination of the
previous tweak with the not scanned outputs.

The code in this change uses the tweak with all the outputs to check
before continuing with the following tweak if at least one spout is
found with the former.