feat(encoding): add SilentPaymentCode::get_placeholder_p2tr_spk
The SilentPaymentCode::get_placeholder_p2tr_spk will be useful while
creating a silent payment transaction to create placeholder outputs to
place in the transaction before derivation, and then coin select the the
inputs at the same time outputs are easily matched with the belonging
silent payment code for later replacement after derivation.

The use of the hashed amount is to easily find the silent payment code
and the desired amount to send. But this may be optimized in the future
for the use of (ScriptBuf, Amount) tuples as keys in HashMaps, whatever
solution performs best.