feat(scan): add Scanner::get_spks_from_tweak
There are setups that doesn't have access to the full transaction at the
moment of scanning for silent payment outputs, less to the prevouts of
that transaction. To facilitate the scanning for those cases, the
computation of the tweaks used for the shared secret derivation can be
outsourced to server, and request them on demand, to get the script
pubkeys resulted from those tweaks.

This change adds a method to derive this silent payment script pubkeys
for the base spending public key and all the associated labels.